+++ src/data/types.ts
@@ -17,3 +17,5 @@ export type VolunteerRow = Tables['volunteers']['Row'];
 export type VolunteerID = VolunteerRow['id'];
 export type Response = Database['public']['Enums']['invited'];
 export type TransactionRow = Tables['transactions']['Row'];
+export type TokenID = Tables['tokens']['Row']['id'];
+export type TransactionStatus = Database['public']['Enums']['transaction_status'];