+++ src/lib/components/Transactions.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { TransactionRow } from '$data/types';
+	import type { TransactionRow, VenueID } from '$data/types';
 	import Feedback from './Feedback.svelte';
 	import ScholarLink from './ScholarLink.svelte';
 	import Table from './Table.svelte';
@@ -8,27 +8,38 @@
 	import VenueLink from './VenueLink.svelte';
 
 	let {
-		transactions
+		transactions,
+		venues
 	}: {
 		transactions: TransactionRow[];
+		venues: { id: VenueID; title: string }[];
 	} = $props();
 </script>
 
 {#snippet row(transaction: TransactionRow)}
 	<tr>
-		<td><Tag>{transaction.purpose}</Tag></td>
-		<td><Tokens amount={transaction.tokens.length} /></td>
+		<td
+			>{#if transaction.tokens === null}unknown{:else}<Tokens
+					amount={transaction.tokens.length}
+				/>{/if}</td
+		>
 		<td
 			>{#if transaction.from_scholar}<ScholarLink
 					id={transaction.from_scholar}
-				/>{:else if transaction.from_venue}<VenueLink id={transaction.from_venue}
+				/>{:else if transaction.from_venue}<VenueLink
+					id={transaction.from_venue}
+					name={venues.find((v) => v.id === transaction.from_venue)?.title ?? 'unknkown venue'}
 				></VenueLink>{/if}</td
 		>
 		<td
 			>{#if transaction.to_scholar}<ScholarLink
 					id={transaction.to_scholar}
-				/>{:else if transaction.to_venue}<VenueLink id={transaction.to_venue}></VenueLink>{/if}</td
+				/>{:else if transaction.to_venue}<VenueLink
+					id={transaction.to_venue}
+					name={venues.find((v) => v.id === transaction.to_venue)?.title ?? 'unknkown venue'}
+				></VenueLink>{/if}</td
 		>
+		<td>{transaction.purpose}</td>
 	</tr>
 {/snippet}
 
@@ -37,15 +48,13 @@
 {:else}
 	<Table>
 		<tr>
-			<th>Purpose</th>
 			<th>Tokens</th>
 			<th>From</th>
 			<th>To</th>
+			<th>Purpose</th>
 		</tr>
 		{#each transactions.sort((a, b) => new Date(b.created).getTime() - new Date(a.created).getTime()) as transaction}
-			<tr>
-				{@render row(transaction)}
-			</tr>
+			{@render row(transaction)}
 		{/each}
 	</Table>
 {/if}