+++ src/lib/components/VenueLink.svelte
@@ -4,7 +4,7 @@
 
 	interface Props {
 		id: string;
-		name?: string;
+		name: string;
 	}
 
 	let { id, name }: Props = $props();
@@ -12,18 +12,4 @@
 	const db = getDB();
 </script>
 
-{#snippet link(id: string, name: string)}
-	<Link to="/venue/{id}">{name}</Link>
-{/snippet}
-
-{#if name === undefined}
-	{#await db.getSource(id)}
-		...
-	{:then source}
-		{@render link(id, source.name)}
-	{:catch}
-		?
-	{/await}
-{:else}
-	{@render link(id, name)}
-{/if}
+<Link to="/venue/{id}">{name}</Link>