+++ src/routes/currency/[id]/transactions/+page.ts
@@ -1,3 +1,4 @@
+import getVenueTitles from '$lib/data/venueTitles';
 import type { PageLoad } from './$types';
 
 export const load: PageLoad = async ({ parent, params }) => {
@@ -10,7 +11,14 @@ export const load: PageLoad = async ({ parent, params }) => {
 		.or(`from_venue.eq.${params.id},to_venue.eq.${params.id}`);
 	if (transactionsError) console.log(transactionsError);
 
+	const { data: venues, error: venueError } =
+		transactions === null
+			? { data: null, error: null }
+			: await getVenueTitles(supabase, transactions);
+	if (venueError) console.log(venueError);
+
 	return {
-		transactions
+		transactions,
+		venues
 	};
 };